+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -128,7 +128,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdw0p5pc4sny6s95kh84244q"),
+  _id: z.string().default("conv_01kdwqh4ervq8mzdz4abycdtjq"),
   entityId: z.string(),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
@@ -161,7 +161,7 @@ export const ConversationResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationResponse
 > = z.object({
-  id: z.string().default("conv_01kdw0p5pc4sny6s95kh84244q"),
+  id: z.string().default("conv_01kdwqh4ervq8mzdz4abycdtjq"),
   entityId: z.string(),
   kind: ConversationResponseKind$outboundSchema,
   displayName: z.string(),