+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -307,7 +307,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdw0p5pkwz5ay0y32cmb0hrs"),
+  _id: z.string().default("conv_01kdwqh4f051sga852qennxm22"),
   entityId: z.string(),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
@@ -341,7 +341,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kdw0p5pkwz5ay0y32cmb0hrs"),
+  id: z.string().default("conv_01kdwqh4f051sga852qennxm22"),
   entityId: z.string(),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),